objId is size_t and compulsory
git-svn-id: file:///home/roberto/Git/svn@7151 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 71711c3a6b626e0f253a3098a241df651fd304b6

+++ trunk/xmipp/applications/programs/scale/scale_main.cpp
@@ -77,7 +77,7 @@ protected:
             nThreads = getIntParam("-interp", "fourier");
     }
 
-    void processImage(const FileName &fnImg, const FileName &fnImgOut, long int objId)
+    void processImage(const FileName &fnImg, const FileName &fnImgOut, size_t objId)
     {
         Image<double> img;
         img.readApplyGeo(fnImg,mdIn,objId);
